chore: add constants file to `common` package starting with mnemonic
