+++ packages/common/src/constants.ts
@@ -1,0 +1,14 @@
+export const mnemonic = [
+  "habit",
+  "hope",
+  "tip",
+  "crystal",
+  "because",
+  "grunt",
+  "nation",
+  "idea",
+  "electric",
+  "witness",
+  "alert",
+  "like",
+]
